Fix error on loading icons and change components splitting code with lazy function
