+++ src/helpers/default-album-cover.png
