+++ src/roles/host/sagas/tasks/spotify/searchTracks.spec.js
@@ -1,5 +1,5 @@
 import { call } from 'redux-saga/effects'
-import getSpotifyClient from '../spotify/getSpotifyClient'
+import getSpotifyClient from './getSpotifyClient'
 import searchTracks from './searchTracks'
 
 test('Returns the results from the Spotify API', () => {
