+++ src/roles/oauth/App.js
@@ -1,31 +1,24 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import qs from 'querystring'
-import SpotifyApiFactory from 'helpers/spotifyApi'
 import { Redirect } from 'react-router-dom'
 import { formatName } from 'helpers/formatName'
+import { v4 as uuid } from 'uuid'
 
 export default function App ({ history, location }) {
   const hash = location.hash.substr(1)
   const params = new window.URLSearchParams(hash)
 
   const accessToken = params.get('access_token')
-  const API = SpotifyApiFactory({accessToken});
 
-  async function getProfile(){
-    const profile = await API.profile.getData();
-
-
-    const hostParams = {
-      party: formatName(profile.display_name),
-      code: Math.floor(100000 + Math.random() * 900000),
-      accessToken
-    }
-
-    return <Redirect to={`/host?${qs.stringify(hostParams)}`} />
+  const hostParams = {
+    party: formatName(uuid()),
+    code: Math.floor(100000 + Math.random() * 900000),
+    accessToken
   }
 
-  getProfile();
+  return <Redirect to={`/host?${qs.stringify(hostParams)}`} />
+
 }
 
 App.propTypes = {
